fixed the problem with plagas y enfermedades modules

+++ database/migrations/2020_09_02_084905_create_new_suelo_table.php
@@ -18,7 +18,7 @@ class CreateNewSueloTable extends Migration
             $table->string('parcela_id');
             $table->string('materia_organica');
             $table->string('textura');
-            $table->integer('pH');
+            $table->decimal('pH');
             $table->unsignedBigInteger('submission_id');
             $table->timestamps();
         });
